requirements.txt
google-cloud-bigquery==3.13.0
pandas==2.0.3
python-dotenv==1.0.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.35.0
tokenizers==0.14.1
scikit-learn==1.3.2
numpy==1.24.4
matplotlib==3.7.4